Use index_fd() in generate_submodule_summary()
Use the index_fd() function in generate_submodule_summary() to have a
less error-prone and more concise set of operations.
-----8<-----
Quote from Dscho:
> +			 */
> +			struct child_process cp_rev_parse = CHILD_PROCESS_INIT;
> +			struct strbuf sb_rev_parse = STRBUF_INIT;
> +
> +			cp_rev_parse.git_cmd = 1;
> +			cp_rev_parse.no_stderr = 1;
> +			cp_rev_parse.dir = p->sm_path;
> +			prepare_submodule_repo_env(&cp_rev_parse.env_array);
> +
> +			argv_array_pushl(&cp_rev_parse.args, "rev-parse",
> +					 "HEAD", NULL);
> +			if (!capture_command(&cp_rev_parse, &sb_rev_parse, 0)) {
> +				strbuf_strip_suffix(&sb_rev_parse, "\n");
> +				get_oid_hex(sb_rev_parse.buf, &p->oid_dst);
> +			}
> +			strbuf_release(&sb_rev_parse);
> +		} else if (S_ISLNK(p->mod_dst) || S_ISREG(p->mod_dst)) {
> +			struct child_process cp_hash_object = CHILD_PROCESS_INIT;
> +			struct strbuf sb_hash_object = STRBUF_INIT;
> +
> +			cp_hash_object.git_cmd = 1;
> +			argv_array_pushl(&cp_hash_object.args, "hash-object",
> +					 p->sm_path, NULL);
> +			if (!capture_command(&cp_hash_object,
> +					     &sb_hash_object, 0)) {
> +				strbuf_strip_suffix(&sb_hash_object, "\n");
> +				get_oid_hex(sb_hash_object.buf, &p->oid_dst);
> +			}
> +			strbuf_release(&sb_hash_object);

It would probably be shorter, less error-prone, and quicker to use
`index_fd()` directly.
----->8-----

The statement Christian gave had to be modified a bit.
-----8<-----
Quote from Christian:
I haven't tried it but it looks like using index_fd() to hash a file
might be done like this:

               struct stat st;
               struct object_id oid;
               int fd = open(path, O_RDONLY);

               if (fd < 0 || fstat(fd, &st) < 0 || index_fd(&the_index,
                                                            &oid, fd, &st, 2,
                                                             prefix, 3))
                    die(_("Couldn't hash object from '%s'), path);

I am not sure we should die() in case of error, but hopefully you get the idea.

Please try it on a new branch and send the branch to us (whether it
works or not).
----->8-----

I had to do a 'oidcpy()' to make sure the destination OID gets stored
in case of a success of the above if-condition. Also, as Christian
commented, I had to error() out instead of die()ing out.

I built this branch over 'summary-v2-issue-7' so that the 'git log'
failure does not crawl up and it is easier to squash it later.

All tests pass.

Sources:
https://www.ibm.com/support/knowledgecenter/en/ssw_ibm_i_74/apis/fstat.htm
https://pubs.opengroup.org/onlinepubs/009695399/functions/fstat.html