Create test for checking output of deinit SMs
Create a test for checking the summary output of deinit SMs.
Quoting Kaartic:
-----8<-----
No. What I had in mind is a test that ensures the 'summary' sub-command
does not print anything when a submodule is de-initialized. It should
also initialize a submodule and ensure that we get a proper summary for
it after the initialization.
----->8-----

I think the SM is not deinitialsing for some reason. And if I try to
init it again, I am met with the error that the url 'sm1' does not
exist in the .gitmodules.

This is just a placeholder commit. I will make a better commit after I
sort this out.