t7401: issue a warning and a NEEDSWORK regarding test scipt
Issue a warning regarding the usage of 'git add' instead of 'git
submodule add' to add submodules to the superproject. Also issue a
NEEDSWORK regarding the outdated syntax and working of the test, which
may need to be improved to obtain better and desired results. Change the
'test' to 'test script' in the test description for less ambiguity.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>