submodule: add quiet option
The quiet option was never really taken care of in the C version. Add
a rudimentary implementation of it to make sure it works.

To verify my workings, I added a 'quiet' in t7421.2 when computing the
summary. Since 'quiet' suppresses everything except the errors, the
test_cmp should have failed since the actual output should have failed,
but it was opposite:
---8<---
BEFORE THIS COMMIT, output of t7401.2:

git submodule summary -q HEAD^
+ git -C sm rev-parse --short HEAD^
+ rev1=35b40f1
+ git -C sm rev-parse --short HEAD
+ rev2=dbd5fc8
+ cat
+ test_cmp expected actual
+ eval diff -u "$@"
+ diff -u expected actual
ok 2 - submodule summary output for initialized submodule
--->8---

AFTER THIS COMMIT, output of t7401.2:

+ git submodule summary -q HEAD^
+ git -C sm rev-parse --short HEAD^
+ rev1=35b40f1
+ git -C sm rev-parse --short HEAD
+ rev2=dbd5fc8
+ cat
+ test_cmp expected actual
+ eval diff -u "$@"
+ diff -u expected actual
--- expected    2020-07-27 17:36:02.584814162 +0000
+++ actual      2020-07-27 17:36:02.580813794 +0000
@@ -1,3 +1 @@
-* my-subm 35b40f1...dbd5fc8 (1):
-  > add file2

error: last command exited with $?=1
not ok 2 - submodule summary output for initialized submodule
------
This thus shows that this patch does have the effect which was desired
though the implementation can be beautified more. Please help in that
regard. If this patch is not necessary then I will obviously drop it.