submodule: document the '--for-status' option
The 'for-status' option is used to compute the summary of submodule(s)
in a superproject by skipping the ignored submdules i.e., those with
'ignore_config' set to 'all' in the '.gitmodules' or '.git/config'.

The option was introduced in d0f64dd44d (git-submodule summary:
--for-status option, 2008-04-12) and was refined in
3ba7407b8b (submodule summary: ignore --for-status option, 2013-09-06)
but was not mentioned in the 'git submodule' Documentation.

Document the '--for-status' option used in the command 'git submodule
summary'.`

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>