t7421: eliminate 'grep' check in t7421.4 for mingw compatibility
The 'grep' check in test 4 of t7421 resulted in the failure of t7421 on
Windows due to a different error message

    error: cannot spawn git: No such file or directory

instead of

    fatal: exec 'rev-parse': cd to 'my-subm' failed: No such file or directory

Hence, eliminate the 'grep' check in t7421. Instead, verify the
presence of an error message by doing a 'test_must_be_empty' on the
file containing the error.

Also, tighten up the check to compute '{src,dst}_abbrev' by guarding the
'verify_submodule_committish()' call using 'p->status !='D'', so that
the former isn't called in case of non-existent submodule directory.

Reported-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Helped-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Signed-off-by: Shourya Shukla <shouryashukla.oo@gmail.com>